backend/requirements.txt
fastapi
pandas
uvicorn
python-multipart
matplotlib
seaborn
google-generativeai
python-dotenv
